refactor: keep FFmpeg constants with its Worker
