+++ apps/ffmpeg-mcp/worker.js
@@ -1,14 +1,14 @@
 import { createMcpHandler, McpServer } from "@modelcontextprotocol/server";
 import { z } from "zod";
+import { MCP_USAGE_HEADERS } from "../../shared/registry/mcp.ts";
+import { validateUserMediaUrl } from "../../shared/user-media-url.ts";
 import {
     calculateFfmpegCharge,
     FFMPEG_COST_PER_SECOND,
     FFMPEG_MAX_MEDIA_BYTES,
     FFMPEG_MAX_RUN_MS,
     FFMPEG_OUTPUT_EXTENSIONS,
-} from "../../shared/ffmpeg.ts";
-import { MCP_USAGE_HEADERS } from "../../shared/registry/mcp.ts";
-import { validateUserMediaUrl } from "../../shared/user-media-url.ts";
+} from "./ffmpeg.js";
 
 const MAX_SOURCE_REDIRECTS = 5;
 const ADJUSTMENT_ID = "cloudflare.container.basic_runtime.v1";
